Further pep8 refactor
