+++ Luna/__init__.py
@@ -1,4 +1,4 @@
-from Luna.core.loggingFn import Logger
+from Luna.core.loggers import Logger
 
 
 __version__ = "0.0.1"
