+++ Luna/core/exceptions.py
@@ -1,4 +1,4 @@
-from Luna.core.loggingFn import Logger
+from Luna.core.loggers import Logger
 
 
 class ProjectNotSet(Exception):
