+++ Luna/workspace/build.py
@@ -1,1 +1,1 @@
-from Luna.core.loggingFn import Logger
+from Luna.core.loggers import Logger
