+++ configs/Luna_config.json
@@ -1,6 +1,8 @@
 {
+    "project.recent": [],
+    "python.commandPort": 7221,
     "logging.level": 10,
     "hud.block": 5,
-    "hud.section": 7,
-    "general.commandPort": 7221
+    "project.previous": "",
+    "hud.section": 7
 }
