+++ tests/logging_test.py
@@ -1,4 +1,4 @@
-from Luna.core.loggingFn import Logger
+from Luna.core.loggers import Logger
 
 
 Logger.info("Test info")
